refactor: 엔티티 파일명 modules에서 models로 변경
