+++ alembic/env.py
@@ -11,7 +11,7 @@ from sqlmodel import SQLModel
 
 from app.core.config import settings
 
-from app.modules.skill.modules import Skill
+from app.modules.skill.models import Skill
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
