+++ app/modules/member/models.py
@@ -7,8 +7,8 @@ from sqlmodel import Field, Relationship
 from app.shared.models.base import BaseModel
 
 if TYPE_CHECKING:
-    from app.modules.resource.modules import Resource
-    from app.modules.favorite.modules import Favorite
+    from app.modules.resource.models import Resource
+    from app.modules.favorite.models import Favorite
 
 class Member(BaseModel, table=True):
     __tablename__ = "members"
