+++ app/modules/notice/models.py
@@ -6,18 +6,18 @@ from sqlmodel import Field, Relationship
 from app.shared.models.base import BaseModel
 
 if TYPE_CHECKING:
-    from app.modules.project.modules import Project
+    from app.modules.project.models import Project
 
-class Team(BaseModel, table=True):
-    __tablename__ = "teams"
+class Notice(BaseModel, table=True):
+    __tablename__ = "notices"
 
-    project: "Project" = Relationship(back_populates="teams")
+    project: "Project" = Relationship(back_populates="notices")
 
     id: UUID = Field(
         default_factory=uuid4,
         primary_key=True,
         nullable=False,
-        description="팀 고유키"
+        description="공지 고유키"
     )
 
     project_id: UUID = Field(
@@ -25,13 +25,12 @@ class Team(BaseModel, table=True):
         description="프로젝트 고유키"
     )
 
-    name: str = Field(
+    title: str = Field(
         nullable=False,
-        description="팀 이름"
+        description="공지 제목"
     )
 
-    detail: str | None = Field(
-        default=None,
-        nullable=True,
-        description="팀 설명"
+    detail: str = Field(
+        nullable=False,
+        description="공지 내용"
     )
