+++ app/modules/project/models.py
@@ -7,10 +7,10 @@ from sqlmodel import Field, Relationship
 from app.shared.models.base import BaseModel
 
 if TYPE_CHECKING:
-    from app.modules.team.modules import Team
-    from app.modules.resource.modules import Resource
-    from app.modules.notice.modules import Notice
-    from app.modules.work.modules import Work
+    from app.modules.team.models import Team
+    from app.modules.resource.models import Resource
+    from app.modules.notice.models import Notice
+    from app.modules.work.models import Work
 
 class Project(BaseModel, table=True):
     __tablename__ = "projects"
