+++ app/modules/skill/repository.py
@@ -5,7 +5,7 @@ from sqlalchemy import func
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlmodel import select
 
-from app.modules.skill.modules import Skill
+from app.modules.skill.models import Skill
 
 
 class SkillRepository:
